requirements.txt
# YOLOv9n зависимости для vehicle damage detection
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pillow>=9.0.0
opencv-python>=4.5.0
requests>=2.25.0
aiofiles>=0.8.0
python-multipart>=0.0.6

# LLM анализ - OpenRouter API
httpx>=0.24.0
openai>=1.0.0

# ML модели - YOLOv9n (легковесная)
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy>=1.21.0

# SAM для точной сегментации повреждений
segment-anything>=1.0.0
timm>=0.9.0

# CLIP для семантического анализа
open-clip-torch>=2.20.0
